third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbobjectstore_createIndex3-usable-right-away-expected.txt

This is a testharness.js-based test.
PASS IDBObjectStore.createIndex() - the index is usable right after being made 
Harness: the test ran to completion.
